[FIX] stock_location_bin_name: remove constraint on 'location_name_format'
This constraint is blocking the creation of `stock.location` records
with `location_name_format`, while the children are not yet created, so
the location is at first considered as a bin.

This is blocking the loading of XML data for instance where we create
parent locations used by some configurations, and children are then
created later.