[MIG] stock_move_packaging_qty: Migration to 18.0
Packaging management is handled directly in Odoo core.
The decision was made to reuse most of the core functionality, such as reserved/demanded packages and reports.
The ability to set done packages and propagate done packages between linked stock move lines has been migrated.
The ability to modify the reserved quantity when the picking is unlocked and the package is changed has been eliminated, as it makes no sense for a package change to modify customer data. Furthermore, the ability to send more quantities or more packages directly in the core is correctly managed.
A new functionality has been added in an installation to copy computation of reserved packages to completed packages to fill in past data.